python-restrictedpython: fix downloading during do_compile
During do_compile if setup.py finds that pytest-runner - which is
declared as a setup dependency in setup.py - isn't available it was
attempting to download it from pypi, which is problematic if you are
behind a proxy because the environment for do_compile isn't set up for
that. Add an appropriate native dependency to prevent this situation.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>