python3-gtts: switch to using pypi instead of custom wheel class
The custom class was introduced in commit db5a31b71d9a ("introduce
python3-whl.bbclass to be able to directly build Python wheels when a
.tar.gz is not available") to work-around unstable github archives.

Instead of that custom code, we can simply use existing infrastructure
and use the git SHA1 which is stable. Doing so avoids having to carry
any extra custom code, and keeps this recipe in line with various other
recipes which are doing the same.

Signed-off-by: André Draszik <git@andred.net>